Move plotting to dynamics model parent class
This commit moves plotting to the dynamics model parent class
